backend/pyproject.toml
[project]
name = "app-backend"
version = "0.1.0"
description = "FastAPI + SQLModel backend para el stack hibrido Astro/React."
authors = [{ name = "Full-Stack Template" }]
requires-python = ">=3.13.11"
dependencies = [
  "fastapi[standard]>=0.115.5",
  "uvicorn>=0.34.0",
  "sqlmodel>=0.0.24",
  "pydantic-settings>=2.7.1",
  "alembic>=1.14.0",
  "asyncpg>=0.30.0",
  "aiosqlite>=0.20.0",
  "bcrypt==4.1.3",
  "websockets>=13.1",
  "orjson>=3.10.0",
  "pyjwt>=2.9.0",
  "passlib[bcrypt]>=1.7.4",
  "pip-audit>=2.9.0",
  "tzdata>=2024.2",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.7.1",
  "mypy>=1.18.0",
  "pytest>=8.3.3",
  "pytest-asyncio>=0.24.0",
  "httpx>=0.27.2",
]

[tool.uv]
package = true

[tool.setuptools.packages.find]
include = ["app"]
exclude = ["alembic*", "tests*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
  "E", "F", "W", "B", "UP", "A", "C4", "I", "TID", "RUF", "SIM", "PTH", "SLOT",
]
ignore = ["B008"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["."]

[tool.mypy]
python_version = "3.11"
disallow_untyped_defs = true
check_untyped_defs = true
no_implicit_optional = true
strict_optional = true
warn_unused_ignores = true
warn_return_any = true
ignore_missing_imports = true
